+++ mmyolo/utils/misc.py
@@ -93,56 +93,6 @@ def get_file_list(source_root: str) -> [list, dict]:
     return source_file_path_list, source_type
 
 
-def show_class_list(classes, class_num):
-    """Print the data of the class obtained by the current run."""
-    print('\n\nThe information obtained is as follows:')
-    class_info = PrettyTable()
-    class_info.title = 'Information of dataset class'
-    # List Print Settings
-    # If the quantity is too large, 25 rows will be displayed in each column
-    if len(classes) < 25:
-        class_info.add_column('Class name', classes)
-        class_info.add_column('Bbox num', class_num)
-    elif len(classes) % 25 != 0 and len(classes) > 25:
-        col_num = int(len(classes) / 25) + 1
-        class_nums = class_num.tolist()
-        class_name_list = list(classes)
-        for i in range(0, (col_num * 25) - len(classes)):
-            class_name_list.append('')
-            class_nums.append('')
-        for i in range(0, len(class_name_list), 25):
-            class_info.add_column('Class name', class_name_list[i:i + 25])
-            class_info.add_column('Bbox num', class_nums[i:i + 25])
-
-    # Align display data to the left
-    class_info.align['Class name'] = 'l'
-    class_info.align['Bbox num'] = 'l'
-    print(class_info)
-
-
-def show_data_list(args, area_rule):
-    """Print run setup information."""
-    print('\n\nPrint current running information:')
-    data_info = PrettyTable()
-    data_info.title = 'Dataset information'
-    # Print the corresponding information according to the settings
-    if args.val_dataset is False:
-        data_info.add_column('Dataset type', ['train_dataset'])
-    elif args.val_dataset is True:
-        data_info.add_column('Dataset type', ['val_dataset'])
-    if args.class_name is None:
-        data_info.add_column('Class name', ['All classes'])
-    else:
-        data_info.add_column('Class name', [args.class_name])
-    if args.func is None:
-        data_info.add_column('Function', ['All function'])
-    else:
-        data_info.add_column('Function', [args.func])
-    data_info.add_column('Area rule', [area_rule])
-
-    print(data_info)
-
-
 def show_data_classes(data_classes):
     """When printing an error, all class names of the dataset."""
     print('\n\nThe name of the class contained in the dataset:')
