+++ requirements/runtime.txt
@@ -1,1 +1,2 @@
 numpy
+prettytable
