+++ requirements/tests.txt
@@ -6,7 +6,6 @@ isort==4.3.21
 kwarray
 memory_profiler
 parameterized
-prettytable
 protobuf<=3.20.1
 psutil
 pytest
